+++ site/news.md
@@ -4,6 +4,21 @@ title: News
 description: News and announcements for all things Bootstrap Table, including new releases.
 ---
 
+## Bootstrap Table 1.15.3
+
+<span class="post-date">11 July 2019</span>
+
+- **New:** Added nl-BE, fr-CH and fr-LU locale.
+- **Update:** Updated nl-NL, pt-BR, fr-BE, fr-FR, nl-BE and nl-NL locale.
+- **Update:** Fixed treegrid duplicate rows bug.
+- **Update:** Fixed `updateCellByUniqueId` method bug on a filtered table.
+- **Update:** Fixed colspan group header display bug.
+- **Update:** Fixed table footer display bug in some case.
+- **Update:** Fixed `getOptions` bug.
+- **Update:** Fixed `detailView` bug when hiding columns.
+- **Update:** Fixed IE minify bug.
+- **Update:** Fixed full screen scrolling bug.
+
 ## Bootstrap Table 1.15.2
 
 <span class="post-date">24 Jun 2019</span>